requirements.txt
torch
torchvision
pillow 
os
matplotlib

